New parameter "nocache"
In some times the media data is periodicaly overwritten. Because of the cache behaviour of browsers this can effect in displaying old versions of the media documents. To avoid this I've added a possibility to switch off the cache by command in the filterlist syntax:
  {{filterlist>[path/files]&nocache=1}}
Default is ''nocache=0'' to enable caching.